# Conflicts:
#	src/gash/router/server/CommandHandler.java
#	src/global/Constants.java

+++ src/global/Constants.java
@@ -2,18 +2,20 @@ package global;
 
 public class Constants {
 
-	public static int sizeOfChunk = 1024 * 1024 * 1;
+	public static double sizeOfChunk = 1024 * 1024 * 1;
 	public static String dataDir = "C:\\Gossamer\\";
 	public static String clientDir = "C:\\ClientStuff\\";
-	
+
 	public static String jedis1 = "169.254.214.175";
 	public static String jedis2 = "169.254.5.213";
 	public static String jedis3 = "169.254.80.87";
+
 	
 	//public static String jedis1 = "10.250.129.206";
 	//public static String jedis2 = "10.250.47.208";
 	//public static String jedis3 = "10.250.175.205";
 	public static String next="169.254.214.175";
+
 	public static int redisPort = 6379;
 	public static int workPort = 4568;
 	public static String localhost = "127.0.0.1";
